feat: collider shape support setRotation
