+++ packages/design/src/physics/shape/IPlaneColliderShape.ts
@@ -1,13 +1,6 @@
 import { IColliderShape } from "./IColliderShape";
-import { Vector3 } from "@oasis-engine/math";
 
 /**
  * Interface of physics plane collider shape.
  */
-export interface IPlaneColliderShape extends IColliderShape {
-  /**
-   * Set local rotation.
-   * @param normal - The local rotation
-   */
-  setRotation(normal: Vector3): void;
-}
+export interface IPlaneColliderShape extends IColliderShape {}
